requirements.txt
opencv-python
ImageHash
numpy
Pillow
